refactor(lockfile): split format io
